Add files via upload
Created as a project for a university module. Written in C++, the aim was to create a console application that allows the user to play games of noughts and crosses with an AI. The AI was only meant to place random markers, as per the module requirements. However, I added an algorithm (albeit quite basic) to look for potential winning and/or losing places BEFORE placing the 'random' marker. Worked great!