Monopoly Assignment
Created for a university assignment for the "Component Programming" module in my first year, this is game of monopoly made in Visual Studio. Not the greatest of platforms to create a board game however I made it work. Lets the user pick how many players (2-4) they want to participate. Most aspects of the typical game of monopoly works as expected, including property/station purchases, chance/community chest and jail etc.

This project is marked as complete, although I would like to recreate this using everything I have learned during my studies after this was submitted.